:ok_hand: Remove review steps to simplify UI.
Move revert to draft functionality to awaiting supervisor approval step.
